Add a test that passes a multidimensional array from Fortran to Chapel
Add an extern interface for ISO_Fortran_binding.h using c2chapel + hand
conversion. The interface was generated from Intel's version of
ISO_Fortran_binding.h. Add a test that uses the new extern interface to pass
a multidimensional array from Fortran to Chapel and write values to it in
the Chapel procedure. The test is only run for PrgEnv-intel.

Update the --library-fortran code to handle the Fortran array record
specially and generate an array declaration for it in the Fortran interface
module.