Move printchplenv down to chplenv
Legacy style imports don't work in python 3, so in #2814 we updated our chplenv
scripts to insert $CHPL_HOME/util/chplenv into the $PYTHONPATH. Unfortunately,
this creates other issues in pyton 3 where modules are now available under 2
different names, so we end up rerunning scripts twice. This results in things
like double warnings and wasting time recomputing results. This makes tests
like test/chplenv/chplconfig/warnings/ fail, but more importantly results in
poor user experience because warnings are printed multiple times.

To resolve this and maintain python 2 and 3 compatibility this just
moves printchplenv into the chplenv dir and we'll add a bash wrapper to
call it. This means all chplenv scripts are now in the same directory,
so modules are only available under a single name. This move is a little
unfortunate, but this was the cleanest solution I could come up with.