Alter Cython build command on Crays
Cython tries to always use the compiler it was built with when compiling the
Cython source files.  This is usually what you want, but on Crays we want to use
the programming environment wrapper instead of the compiler directly - doing
otherwise leads to failures with certain environment variables settings (e.g.
CHPL_REGEXP=re2).

Because this will alter the compiler used to something that may not be ABI
compatible with how Cython was built, generate a warning when the programming
environment is something other than Gnu.  If the user knows their Cython was
built with something else, they can choose to ignore this warning.