Get comm=ofi network AMOs working.
Until now, although the ofi comm layer supported the network atomics
interface, it could actually only do "network" AMOs by using AMs to
initiate processor AMOs on the node owning the object.  Here, add
support for actual network AMOs.  We now make the appropriate libfabric
calls whenever the target datum is in registered memory and the provider
supports the requested operation on the datum's type.

While here, I also simplified the AMO processing to use a single set of
functions and a single AM request type for all AMOs, fetching and not.
This requires a little bit of extra logic in the functions in order to
avoid manipulating operands or results that aren't present, and will
cause us to transmit an extra word for AM-mediated non-fetching AMOs and
atomic GETs, but the extra logic and/or 8 bytes won't affect operation
rates noticeably and the reduced number of functions seems well worth
it.

While here, I fixed a bug (race) which could cause a hang due to lack of
progress, when the AM handler's failure to check its transmit completion
counter resulted in transmitted 'AM done' indicators never actually
making it onto the network.

There are also some small improvements to the debug output capabilities
here, to allow printing AMO operands in as useful a format as possible.