Increase the size of small forks under ugni
We have two different mechanisms for spawning remote tasks under ugni,
"small" forks and "large" forks. For small forks only a single put is
required, whereas large forks involve a put, followed by a get from the
remote node as well as another fork back to original node to free the
argument bundle. Because of the extra get and AM, small forks are much
faster. The previous threshold for small fork payloads was 64 bytes and
this raises it to 1,024 bytes. Here we trade-off remote task spawning
speed for memory footprint. There is reserved landing space for the
small fork buffers for each remote node. For a 1,024 node job with 64
comm domains per node (we use a comm domain per core up to 128) we used
to require 8 MiB for remote fork landing pads, and now we'll require 128
MiB. A large job (say 8K nodes with 128 comm domains) will require 2
GiB, which is large, but not horrible. The theoretical max for an XC is
32K nodes, which would be 8 GiB, which is excessive and we'll need to
revisit if we start running at that sort of scale.

Today, most people are running at much smaller scales and we believe the
improvement in remote task spawning is worthwhile. The typical argument
bundle for the loops in kernels of our multi-locale benchmarks like
STREAM, RA, PRK-Stencil, SSCA, MiniMD, and others is between 500 and 800
bytes, so 1,024 should allow us to almost always use small forks. At 16
nodes, small forks are 60% faster than large forks and I expect large
forks to scale poorly because of the remote fork to the original node to
free the argument bundle.