Don't change config const's apparent type in compopts or execopts.
The many-to-one test has a config const (runSecs) whose initializer
makes its type default real.  But the test's compopts files were setting
the default value of that config const using an integer constant.  I cut
and pasted those compopts into the command line for a manual compile,
and then was surprised when a subsequent attempt to override the default
with a real value (matching the apparent implied type in the source
file) on the execution command line resulted in an error message about a
type mismatch.  It just didn't occur to me that by giving an integer at
compile time I'd changed not only the config const's default value but
also its type.  Here, change the compopts files to specify the same
default value as before, except as a real instead of an int, so that the
same sequence of events results in less surprise.

While here, also change the execopts file for the arrayTransfer test so
that the value for the same config const there is also real, as implied
by the source file.  This is just for the sake of clarity, since the
execution-time setting doesn't change the config's type the way the
compilation-time setting can.