Alter handling of c_ptr arguments in Python interoperability
Prior to this change, c_ptr arguments would be left mostly unaltered for the
generated Python files.  This was a problem, because Python doesn't have a
notion of a pointer type - when trying to compile the generated Python files, we
would encounter failures because the type wasn't recognizable to Python.

Here, I've taken the approach that a pointer argument for an exported function
should be treated like a numpy array argument containing that type in Python,
but instead of copying the memory, it should be passed in directly.  It is
assumed that the user will provide the size information themselves or know it
already.

This strategy is needed to support returning c_strings whose memory has been
allocated, as we cannot easily detect this from the code (and so leak memory if
it is returned directly - the return value is copied into a Python bytes string
and the value handled within the generated .pyx, so without the ability to
detect if the memory must be explicitly freed the reference will go out of scope
without having been cleaned up).  Now, the string contents can be copied into
memory provided to the function via the pointer argument.