Add a test locking in the pointer behavior, demonstrating the string arg return
This ensures a program with a c_ptr argument behaves as expected, and mocks up
the intended strategy for returning c_strings that came from Chapel strings (as
opposed to c_strings that were always c_strings).

In summary: create an array in Python to store the information you want to get
back as individual characters.  Pass a pointer to this array and the size
allocated for it to the function.  In the function, memcpy the contents of the
string into this array and return the size used.  Once the function has
returned, remove the excess space from the array, and convert the characters
from their character code back into a string.