Expand our remote task spawning microbenchmark
Previously we only tested an emtpy coforall+on, but that's not really a good
representation for how we usually spawn remote tasks. This adds versions that
have a 512 byte and 4096 byte argbundle payload. 512 bytes represents the
typical argument bundle size (at least for our multi-locale benchmarks like
STREAM, RA, PRK-Stencil, SSCA and MiniMD) and 4096 is currently massive. The
largest in our multi-locale performance suite was just over 800 bytes.

This arg bundle size is important, because our runtime does different things
for "large" vs "small" forks, so it's worth having timings for different
argument bundle sizes since the performance will be different. Currently ugni
has a 64 byte threshold for small forks. Small forks only require a single put,
whereas large forks require a put, and then the target does a remote get and an
AM to free the arg on the original node. At 16 nodes, large forks are ~60%
slower than small forks (and likely to scale worse because of the AM to to the
spawning node.)