Refactor the var declaration statement in the parser
This performs a refactoring on the rule in the parser that
handles vars, consts, params, refs, and const refs to reduce
the amount of duplicated code.  It does this by moving the
choice of keywords into a new rule that converts the keywords
into the appropriate flag or flags (if any).  This replaces
the parser-level variables that were used for this purpose
previously.