Update c2chapel to generate two overloads for functions with pointer arguments
For a function with a pointer argument, c2chapel generates a declaration with
a `ref` intent argument.

int f(int* x); => extern proc f(ref x: int): int;

This doesn't resolve for an actual that is a `c_ptr(int)`, such as a pointer
returned from another extern function.  In addition to the above, generate
an overload that uses `c_ptr(*)` arguments:

extern proc f(x: c_ptr(int)): int;

This way we get both the `ref` behavior, and the ability to work directly with
`c_ptr`s.