Update a missed call to buildVarDecls
Somehow, my linux64 build on chapcs00 built even though I was passing
a set<flag> in where a set<flag>* was expected (?!). Travis caught it,
though.

This updates the signature in question by passing in a NULL instead
(it didn't really want to create and pass in a seg<flag> anyway) and
guarding the logic that operates on and frees the flags to check
for NULL.