Update some buffer sizes in launch-slurm-gasnetrun_ibv.c
GCC 8.2 complains that these sprintf calls could overflow, which is accurate.
This PR takes the very simple step of increasing the buffer sizes for the
destination buffers. I don't think this code should be calling sprintf at all
(snprintf is a better idea in C programs) but I also think the whole thing
needs to be rewritten in Python so I didn't try to go further.