Rework transmit/receive context counting and assignment.
For the most part this change is just to clarify the code.  We have
been blocking the transmit contexts in the per-transmit-info table
(ptiTab) into an initial chunk used by worker threads and a trailing
chunk used by the AM handler(s).  This blocking remains.  However,
here we rearrange the code to delay computing how many worker transmit
contexts we need until right before we create them, and while doing
that computation we figure out if we will be using a fixed or dynamic
assignment of transmit contexts to worker threads.  However, we don't
actually implement that dynamic assignment yet; the old code that just
aborts if we run out of transmit contexts is still there for now.