Suppress gcc 8 spurious array-bounds warnings -> errors (#11781)
Suppress gcc 8 spurious array-bounds warnings -> errors (#11781)

[reviewed by @ronawho ]

Gcc 8 added new array bounds checking that gets confused in complicated code, such as when we use --baseline. Specifically, baseline code is too complicated for gcc 8 to recognize that the result of safeCast() is safe.

This change disables the problematic checks in gcc 8 for generated code, to allow baseline to work.