Rename ptiTab to tciTab, more accurately reflecting its purpose.
The transmit context table was called 'ptiTab' ('pti' for 'per-transmit
info').  Change 'pti' to 'tci' ('transmit context info') to match up
with our other uses of 'tci', notably the common variable 'tcip' for the
pointer to the table entry we have allocated.  This is just a renaming;
there are no functional changes.