Support sharing transmit contexts among multiple worker threads.
In the past the ofi comm layer has bound transmit contexts to both AM
handler threads and worker threads for the duration of execution.  We
made a best guess as to how many worker threads would be present and
then bound contexts to AM handler threads as those started, and to
worker threads at the time of their first transmit.  If we couldn't find
a free context, we took an error exit.

Here, implement sharing of contexts among worker threads when needed.
The AM handler thread still gets its own context.  And if the tasking
layer says that it uses a fixed number of threads, and we can get at
least enough transmit contexts for the AM handler and that many worker
threads, we still bind contexts to threads throughout execution.  But if
either the tasking layer doesn't use a fixed number of threads (example:
fifo) or the ofi provider does not support as many transmit contexts as
the fixed number of threads (example: sockets, with qthreads, on a
compute node with more than 15 cores), now threads will dynamically
acquire and release transmit contexts as they need them.

Up until now we had to set CHPL_RT_NUM_THREADS_PER_LOCALE to some value
less than the provider's maximum number of transmit contexts (16, for
sockets) in order to run tests.  As of this commit that is no longer
necessary -- I tested the multilocale suite with the sockets provider on
a Cray XC system and all the same tests passed as do with comm=ugni.  We
still cannot use the ofi comm layer with tasks=fifo, but that's due to
memory registration issues, not transmit context issue.  Specifically,
we get internal errors for unregistered AM arg bundles because we don't
(yet) duplicate unregistered AM args into bounce buffers, but tasks=fifo
runs the main task on the process rather than a pthread and we don't
register the process stack.  That's easily fixable, but I'm going to do
it separately from this transmit context work.