Snapshot - commented out attempt to resolve issue with type resolution stuff
We're getting an error "extern procedure has arguments of type string" because
the c_sizeof call in dsiBuildArray is getting resolved with strings even though
we're not actually building extern arrays with strings.  We need to limit the
types that portion could get resolved with to a reasonable amount - I think
the partial function I've sketched out will help with that but it needs some
compiler support to detect if a type is an extern type.