Don't try to negate the most-negative signed integer value.
Negating the most-negative value of a C signed integer is undefined
behavior according to the C standard.  The positive value corresponding
to the most negative value does not exist in that type (6.2.6.2 (2)), so
the negation overflows.  The result of integer overflow is not discussed
in the standard, thus it is undefined behavior (4 (2)).

Here, stop negating the most-negative signed integer value when
arranging for signed atomic subtract operations to be done by adding the
corresponding negative value.  The bit-result will be the same on the
two's complement systems we target, so there is no change in observable
behavior.