Only skip multilocale/engin for comm=none.
It works as well with comm=ofi as it does with comm=ugni.  (It doesn't
pass with comm=ofi but it fails for the same superficial reason as it
does with comm=ugni, namely a slight message difference compared to the
output with comm=gasnet.  This can be fixed later.)