Don't force vars onto the heap with comm=ofi.
Comm=ofi can communicate non-heap vars as well as comm=ugni can.

This removes a compilation failure for hello2-module with comm=ofi and
tasks=fifo, where ChapelSerializedBroadcast.chpl:chpl__broadcastGlobal()
and chpl__destroyBroadcastedGlobal() disagreed with the emitted code as
to whether the first argument to those functions was a string* or a
chpl____wide_heap_string*.  It's possible this disagreement reflects a
deeper issue in parallel.cpp:needHeapVars(), but if so I'll leave that
to someone more qualified, because it definitely seems that comm=ofi and
comm=ugni should be treated in the same way here.