Add an explicit error if user tries to use psm or ofi substrates
GASNet-EX doesn't currently support psm or ofi. GASNet will error about
this during it's build, but I wanted to add an explicit error so users
will know to let us know if this is a config they care about.

If this is an important config for somebody we can provide them with
instructions for using a GASNet-1 tarball or something.