Remove user facing documentation for psm and ofi substrates
Add a note to the omnipath (psm) README that it's not currently
supported. Note that I left the gasnetrun_{psm, ofi} launchers and logic
for choosing them in chplenv alone for now. I wanted to have the ability
to send a user a GASNet-1 tarball if either of these configurations are
important for them.