Decrement the CQ counter when we consume the CQ event in ugni
In #11799, we optimized remote task spawning by making non-blocking
forks truly non-blocking in that we don't wait for the CQ event
associated with the remote ACK before continuing. We only decremented
our CQ counter if our local buffer was full, which resulted in increased
pressure on the CQ counter when there were a lot of outstanding
nb-forks.

This moves the CQ count dec into consume_all_outstanding_cq_events right
after we consume the CQ event. This allows other comm events to
decrement the CQ counter, reducing pressure on the CQ counter.

Historically, the CQ counter decrementing logic was separate from
processing the CQ event associated with it. That was originally to avoid
races with free'ing non-blocking descriptors. See internal chapel-code
640d948 for more info. At that point we acquired a non-blocking
descriptor after we acquired the comm domain, so there was a race where
we thought we'd be able to acquire a nb-descriptor because the CQ
counter wasn't full, but the task could have switched after the CQ
counter decrement but before the NB descriptor free.

That logic was all from when the NB descriptors were attached to the
comm domains themselves, but they were swapped out to distinct data
structures. Additionally chapel-code dbc8ef1 moved when the comm domain
is acquired, so we now acquire it after we have a NB descriptor so there
is no longer any race with decrementing the CQ counter as soon as we
process the CQ event.