Allow underscores in the number when casting from string to integral
In the string->int cast function in StringCasts.chpl clean up the string by
stripping leading and trailing whitespace, checking that it has only one word
and removing all underscores unless they're in the first position.

Add a test that ensures we can cast strings with underscores to all of the
integral types.  Add tests that ensure that a leading underscore is an error
and that having multiple words in the string is an error.  If we blindly
removed underscores from the whole string without checking for multiple words
first this last test wouldn't issue the expected error.