Refactor ugni's buffered atomic implementation
This significantly refactors the implementation of buffered atomics in ugni. In
the old world each thread had multiple thread local buffers for each argument
it was buffering and then there were global arrays that were pointers to the
thread local buffers. This resulted in a lot of code duplication and a pretty
hard to follow implementation. It also resulted in a max of 1,024 threads that
could buffer since the global arrays were of fixed size.

Here we replace all that logic with a struct that contains the buffers for each
argument inline. Then the only global information we need is a linked list of
the per-thread structs. This allows for significant cleanup in the
implementation.