Add a helper function to check if the type is appropriate for an extern array
We were running into problems because the compiler was trying to resolve the
extern array code with types that would not work, even though no extern array
was being created with those types as element types.  So add a param function
that will indicate if the type is an appropriate one, which when used will
cause the compiler not to resolve that code with an inappropriate type.

This function will return true if the element type is a primitive type that is
not a string (but could be a c_string), or if the type is declared as an extern
type.  In the future when we can export types, those would also be valid.  To
support the extern type check, I added a compiler primitive.