Add a writeThis override for _channel_regexp_info
For some reason that eludes me, we weren't resolving object.writeThis before
(even though we have tests that call writeln on instances of object) but because
of my changes we are now.  Resolving object's writeThis causes virtual dispatch
to track the writeThis method of all subclasses, which caused us to resolve
_channel_regexp_info.writeThis and fail due to its field "theRegexp" being of
extern type and thus not having a writeThis.  Since it seems reasonable to
resolve object.writeThis, add an explicit override to _channel_regexp_info that
avoids calling writeThis on its fields of extern type.