Add a sizeof_ddata primitive.
The only thing preventing us from calling chpl_mem_array_alloc() rather
than the array_alloc primitive to allocate memory for array data was
that we didn't have a way to get the size of the array element type in C
terms.  More specifically, it was that we didn't have a way to get the
size of a wide pointer, for array elements of class type with no_local.
Simply applying c_sizeof() to the element type always produces the size
of the local class reference pointer, no_local notwithstanding.

Here, add a primitive to get the size of a ddata element type.