Rename sizeof_ddata to sizeof_raw.
As currently implemented, the new primitive for getting the size of
array elements to support array allocation only works for _ddata types
or, more generally, class types.  I'd therefore named it sizeof_ddata.
But eventually I think we'll want to extend its implementation to
support other types and use it for things like remote PUT and GET sizes.
So here, rename it to sizeof_raw.