Fix target-compiler command-line optimizer specialization flag (#11905)
Fix target-compiler command-line optimizer specialization flag (#11905)

[reviewed by @mppf ]

For specializing generated code to a particular CPU architecture, we had been assuming that passing the -march= flag to the target compiler was the usual case, and -mcpu= was specific to ARM. However, it turns out that -mcpu= is the usual case, and march= is specific to x86 architectures. The incorrect assumption was causing problems with architectures we normally are unable to test, such as POWER. This change fixes that problem.

This change also simplifies some of the code by taking advantage of the fact that we are now putting generic architecture types in the Chapel binary paths.

While there, I've removed the last vestiges of x86_64 to aarch64 cross compilation support, which has not worked for a while and which we do not have a way to test. The recent change to depend on generic architecture types being available precludes that kind of cross compilation anyway. (We can cross compile from sandybridge to broadwell, but not x86_64 to aarch64.) With this change, we no longer treat aarch64 as a PLATFORM.

Closes #10905
Closes #11208