Don't test ddata_sizeof_element on 32-bit platforms.
This test is primarily to ensure that _ddata_sizeof_element() produces
what we expect for arrays of classes in multilocale situations, namely
the size of a wide pointer.  That can be adequately tested on 64-bit
systems.  There's no need to go to the trouble of adding more .good
files so that it can pass on 32-bit systems.