Avoid baseline failure involving refs to C function pointers
Under baseline compilation, Chapel was failing at code generation due to turning
off optimizations which converted references to fields into accesses of the
field's contents.  In the case of extern records storing function pointers, this
meant that we were trying to store pointers to function pointers into pointers
to void*, which is not allowed by C compilers.

This is related to #9411