Add two futures to track the failure I encountered with function pointer fields
This tracks the two expected strategies for representing a function pointer
field in Chapel.  One declares it as a c_fn_ptr (which today means we can't do
much with it beyond pass it to another extern function), the other declares it
as a c_void_ptr.  Both fail in the same way when compiled with --baseline (the
Chapel compiler has failed to convert:

```C
_ref_c_void_ptr x = &(instance->c_fn_ptr_field)
```

to

```C
c_void_ptr x = instance.c_fn_ptr_field
```

and so we get an error at C compilation that there is an assignment between
incompatible pointer types.