Remove unnecessary checking for inTree()
This check gave me a pause "why are we checking this? shouldn't it
always be true?". Because we are traversing the ASTs to be normalized,
and I think we do not normalize ASTs that have already been deleted.

Turns out yes, it is always true as far as testing.
Adjusting the code to reflect this understanding.

This is a minor aesthetic change. Not reviewed.