Add a stopgap mpirun launcher for comm=ofi, and out-of-band to match.
Until now we haven't had a way to launch multilocal comm=ofi programs
except on Cray X* systems.  Here, add a very simple launcher based on
OpenMPI mpirun, which allows launching on our chapcs cluster either
within or outside of a slurm job allocation.  (mpirun will auto-sense if
it's running in a slurm allocation and use the slurm-provided resources
if so.)

Related to this, add a comm=ofi out-of-band module that implements the
necessary startup, shutdown, barrier, and allgather operations in terms
of MPI calls.