Catch my cleanup variation up to the fast version in the release
This brings my study version of the pidigits benchmark into better
sync with the version in the release while retaining the "use a single
temp" property.  That said, I moved the temp from a local variable to
an outer-scope variable to avoid creating and destroying it as a
stylistic compromise in hopes that it would help performance.  A
notion of local static variables would help with this...