Fix a bug in llvmAggregateGlobalOps
See also commit 6262b7090d38a7395fe35c024541bf7a43068f47
which introduced a reordering strategy for the memory operations
that are aggregated.

The problem was, the reordering strategy didn't consider memory
operations that were unrelated (such as an unrelated load when
aggregating loads). This could cause use-before-defined errors for the
emitted instruction sequence.

This patch fixes the problem by keeping track of those stores/loads
that are being aggregated in a set and using this set instead of
just checking for instructions that are loads or stores.