Reflecting review feedback: more python coding improvements.
Express the strings to be matched in a more natural way, using
triple-quoted syntax.  Remove the resiliency against changes in the
length of the '---------------' lines; it's likely not needed and is
hard to read.  Put both strings into a single data structure and iterate
over them.  Don't bother with re.compile(); re.sub() can be used without
that.  Finally, don't write to a temporary copy of the .exec.out.tmp and
then rename that afterward.  That's needed if you're writing bash, but
for python it's more natural to read the whole .exec.out.tmp contents,
possibly modify them, and then write them back out into that same file.