A transliteration (and minor cleanup) of the Julia pidigits version
This was an experiment to see whether the second shortest entry in
the pidigits game could be transliterated into Chapel cleanly.
Now that it can be, I'm curious how much faster/slower it is.

I also used printf-style formatting to do our output for a change,
for better or worse.  This was hard to get right, so I added an
n=57 test which checks for a particularly challenging case.