Add a future to track #11816 (get type of generic var field from type variable)
In working on support for opaque pointers to Chapel arrays, I ran into a
situation where I wanted to get the type of the _instance field from an array
type without having to make an instance of the type to get it (as that would
involve creating unnecessary memory usage).  This is not currently possible
today.