Skip static_dynamic test if we need libmpi.
On linux64 platforms the static_dynamic linking test produces linker
errors for its static modes when libmpi is needed, because that is only
available as a shared object.  Skip it in those configurations.  (Note
that this will also skip it on, say, a Cray XC where there actually is a
static MPI library, but we don't run this test on that platform now and
don't seem likely to start doing so.)