Change all dsiLocalSubdomain*() calls to take a required locale argument
My attempt to use Reflection to have ChapelArray issue a nice error
message if a distribution had not been converted over to take a locale
argument in its dsiLocalSubdomain*() calls failed, whether due to a
bug in reflection or my own lack of understanding of how it would
interact with forwarding, so rather than wrestle with it further, I
decided to bite the bullet and change the interface.  This commit
makes the necessary changes to the modules, taking the 'loc' argument
and doing something reasonable with it if that wasn't too hard; or
halting if it was hard and 'loc' was not 'here'.  I added a new
"unimplemented feature" halt to the HaltWrappers module for this
purpose.