Differentiate output using comm==none, rather than comm==gasnet.
These tests produce different output for comm==none and comm!=none.  But
the .good files we had were for comm==gasnet or not.  This worked well
enough, because we don't run these tests with comm=ugni and there were
no other comm layers.  But now we have comm=ofi, so we need to rename
the .good files appropriately.  These tests now pass for all comm
settings.