The mpirun4ofi launcher should be a little more forthcoming.
It's quiet in here.  Yeah, TOO quiet.

We recently added a "-map-by ppr:1:node" option to the mpirun command
line in the mpirun4ofi launcher.  This ensures we don't run more than
one Chapel process per node.  We're having an ongoing discussion about
whether this is desirable in all cases, but in the meantime, we've
observed that it prevents running multilocale programs outside of a
slurm job.  (Because mpirun doesn't allocate nodes itself, something
else has to do so for it.)  That's okay, at least for now, but the
"--quiet" option we added at the same time suppresses the error message
produced when this happens.  So if you inadvertently run a multilocale
program outside a slurm job, it doesn't run and you get no output.
That's user-hostile.

Here, remove the "--quiet" option, and reinstate the prediff filtering
that removes the message we added the "--quiet" to get rid of.