Prevent GASNet builds from triggering a qthread rebuild
In #10390 we started using `./compileline --includes-and-defines` in
order to get include paths and macro defines for qthreads build args.
Under `CHPL_COMM=gasnet`, compileline will fail if gasnet is not built,
so we set it up so that the qthreads build would depend on the gasnet
build. This resulted in building gasnet first, but as an unintentional
side-effect it also meant building different gasnet layers (or switching
from comm none to gasnet) would trigger a qthreads rebuild.

We don't actually need or want any gasnet includes paths/defines so this
updates things to just call compileline with `CHPL_COMM=none` and
removes the qthreads gasnet dependency.