Adjust comm diags-related testing to reflect networkAtomics=none or not.
Testing with comm=ofi produced a number of differences in comm diags
which, in the end, had to do with network atomics support in the comm
layers.  The existing .good files matched the output for comm=gasnet,
which doesn't (yet) support network atomics.  It has never matched the
output for comm=ugni, but we didn't care about that very much because we
weren't running these tests with that comm layer.  When comm=ofi testing
came along that also didn't match, and it failed to match in the same
way as comm=ugni did.  It turned out that the mismatches were due to
executeOnFast operations used for remote AMOs in the networkAtomics=none
configuration.

Here, adjust .good files to differentiate between networkAtomics==none
results and !=none results.  A previous commit added support for this
differentiation in start_test and sub_test; adapt a similar strategy as
used there in the test/distributions/robust/arithmetic/PREDIFF script
that checks results of those tests for the various distributions.