Disable warnings if [op]= overloads' LHS isn't passed by 'ref'
This disables the compiler warning that exists at present if the first
argument of an `[op]=` overload doesn't have `ref` intent.  This
change is motivated by Louis Jenkins' issue #10803 which points out
that for classes this warning isn't useful.  In a follow-up comment on
the issue, I take the argument one step further and say that I'm
not convinced we should be issuing such warnings for any type.

Even if we were to refine the warnings for just some types, doing so
would require new compiler logic after type resolution had taken
place, while the current warning takes place right after parsing, so
it will need to be removed in any case (i.e., even if you think some
cases need warnings, this is probably a necessary first step to
getting there).

I added a new test demonstrating how op= overloads on classes can be
meaningful without 'ref' intents and updated/cloned existing tests
that generated the warning to adapt them for this proposed new world.