Work around a bug in some system <atomic> headers under clang
Clang prefers to use the system <atomic> header when possible. However,
the system headers provided by gcc<5 were incompatible with clang for
atomic_bool. This led to issues building our re2 shim and prevented us
from making `cstdlib` atomics the default under clang (#4536)

To work around this, avoid using the chpl-atomics.h shim and directly
use std::atomic to implement reference counting for the re2 shim. re2
requires std::atomic ops in order to build, so we know this is safe and
we're only using limited atomic ops that do not run into the system
header incompatibility. See #4534 for past attempts to work around this
issue.

This is a little weird, but it's relatively isolated and pretty small.
Our original plan was to use a configure-like check to see if the system
header was broken, but that got messy pretty quickly, and I find this to
be a cleaner solution. A slightly better fix might be to reimplement
reference counting in re2-interface.cc (or maybe use shared_ptr?) and
avoid bringing in C implementation.

This allows us to use cstdlib atomics under clang (and under
clang-included/llvm) when the system gcc is older than 5.0.

Resolves https://github.com/chapel-lang/chapel/issues/11956